% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimaltocomma.R
\name{d2c}
\alias{d2c}
\title{Change all decimal points to commas in the vector provided}
\usage{
d2c(num)
}
\arguments{
\item{num}{a vector of numeric}
}
\value{
a vector of the input with commas replacing the decimal points. This will be a character vector
}
\description{
Change all decimal points to commas in the vector provided
}
\examples{
d2c(c(1, 2.1, 3, 4.123, 5.01, 6, 7.0, 8))
}
