% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosetta_terms}
\alias{rosetta_terms}
\title{rosetta_terms}
\format{An object of class \code{data.frame} with 352 rows and 3 columns.}
\usage{
rosetta_terms
}
\description{
rosetta_terms
}
\keyword{datasets}
