% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{en2fr}
\alias{en2fr}
\alias{fr2en}
\alias{trans}
\title{Title here}
\usage{
en2fr(x, translate = TRUE)

fr2en(x, translate = TRUE)

trans(x, from_french = FALSE)
}
\arguments{
\item{x}{text}

\item{translate}{logical}

\item{from_french}{logical}
}
\description{
Title here
}
