% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{Translate from one language to another}
\usage{
trans(x, to = "english", from = "french", sep = ";")
}
\arguments{
\item{x}{text}

\item{to}{language to translate to}

\item{from}{language to translate from}

\item{sep}{seperator between multiple 'to' languages}
}
\description{
Translate from one language to another
}
