% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{en2fr}
\alias{en2fr}
\alias{fr2en}
\alias{trans}
\title{Translate a word or phrase from one language to another}
\usage{
en2fr(x, translate = TRUE, ...)

fr2en(x, translate = TRUE, ...)

trans(x, from = "french", to = "english", sep = "; ", allow_missing = FALSE)
}
\arguments{
\item{x}{word or phrase to translate}

\item{translate}{Logical. Perform the translation if TRUE.}

\item{...}{arguments passed to \code{\link[=trans]{trans()}}}

\item{from}{language to translate from}

\item{to}{language to translate to}

\item{sep}{seperator between multiple 'to' languages}

\item{allow_missing}{Should the function return the input value if the term
is missing from the dictionary? If \code{FALSE} then the function will issue a
stop statement if anything is missing.}
}
\description{
Translate a word or phrase from one language to another
}
