% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{en2fr}
\alias{en2fr}
\alias{fr2en}
\alias{trans}
\title{Translation between languages}
\usage{
en2fr(x, translate = TRUE, sep = "; ")

fr2en(x, translate = TRUE, sep = "; ")

trans(x, from = "french", to = "english", sep = "; ",
  allow_missing = FALSE)
}
\arguments{
\item{x}{text}

\item{translate}{logical}

\item{sep}{seperator between multiple 'to' languages}

\item{from}{language to translate from}

\item{to}{language to translate to}

\item{allow_missing}{Should the function return the input value if the term
is missing from the dictionary? If \code{FALSE} then the function will issue a
stop statement if anything is missing.}
}
\description{
Translation between languages
}
