% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{en2fr}
\alias{en2fr}
\alias{fr2en}
\alias{trans}
\title{English to French}
\usage{
en2fr(x, translate = TRUE)

fr2en(x, translate = TRUE)

trans(x, to = "english", from = "french")
}
\arguments{
\item{x}{text}

\item{translate}{logical}

\item{to}{language to translate to}

\item{from}{language to translate from}

\item{x}{text}

\item{translate}{logical}

\item{x}{text}
}
\description{
English to French

French to English

Translate from one language to another
}
