% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{en2fr}
\alias{en2fr}
\alias{fr2en}
\alias{trans}
\title{Translation}
\usage{
en2fr(x, translate = TRUE)

fr2en(x, translate = TRUE)

trans(x, from = "french", to = "english", sep = "; ")
}
\arguments{
\item{x}{text}

\item{translate}{logical}

\item{from}{language to translate from}

\item{to}{language to translate to}

\item{sep}{seperator between multiple 'to' languages}
}
\description{
Translation
}
